2012.9.12模板库版本，继承至33版本，修正如35~38中Bug。理论上与38版的区别： 1、添加了模板库功能； 2、部分输出文件末尾保留“\”。 该版本未经大规模测试！！！（修改了部分修改日期，以便上传） modified: cl1cfg.cpp, rca.h(*), dfgraph.cpp, genCL2.cpp, creatCL0.cpp, createInfNew.cpp, dfg2srcNew.cpp(*) added：   libGenContext.cpp, createInterface.cpp

+++ genCL0.cpp
@@ -257,11 +257,11 @@ int RPUConfig::genCL0Context( CL1Config & cl1config, Vector<RCA *> &CL1RCATemp,V
 
 		if(pseudoRCAcounter <= directSSRAMOutPseudoRCANum)
 		{
-			if((*pseudoRCAIter)->rcaSSRAMOutBaseAddr() < pseudoRCASSRAMOutBaseAddr)
+			if(static_cast<unsigned int>((*pseudoRCAIter)->rcaSSRAMOutBaseAddr()) < pseudoRCASSRAMOutBaseAddr)
 				pseudoRCASSRAMOutBaseAddr = (*pseudoRCAIter)->rcaSSRAMOutBaseAddr();
 
 
-			if((*pseudoRCAIter)->rcaSSRAMOutTopAddr() > pseudoRCASSRAMOutTopAddr)
+			if(static_cast<unsigned int>((*pseudoRCAIter)->rcaSSRAMOutTopAddr()) > pseudoRCASSRAMOutTopAddr)
 				pseudoRCASSRAMOutTopAddr = (*pseudoRCAIter)->rcaSSRAMOutTopAddr();
